preferredSize is used if item is a parent and minimumSize if not
#
I was a bit fast and loose with the last commit. I realised afterwards that minimumSize on a parent element doesn’t work the same way it works for an item that is not a parent. and same goes for preferredSize.
Parent item + minimumSize = Children won’t align properly. I don’t get this, because the minimum size definitely sticks.
Not a parent item + preferredSize = Text is squished if size is set too small to fit everything.
So I decided to change it so that export gives parent items preferredSize and non-parent items minimumSize.