Fixed indent guide path to active + added text to (del, import, export) btns
Not incredibly happy with the indent guide path. It's decent and improves it a lot, but there is something that bothers me. I think it's the blue color. I might want to try orange line color or something...

I liked it better when: delete, import and export buttons were just round buttons with icons, but I can't ignore the fact that adding text makes it way clearer.

I always had the fear that people wouldn't realize you can actually generate the code using this... so...