Fixed #14
As I was fixing this I realized this happened because I used to handle updating item properties by clicking an "update" button and just forgot this when I changed the update behavior. It used to simply collected everything in the properties panel and submit then when a button was clicked.

Currently the properties get updated when they are edited.