Line break fix
I fixed the issue #40 by getting rid of some code that ended up being important. So the bug seesaw just tipped the other way. Put on my thinking cap and realized I had made it unnecessarily complicated to begin with, so I cleaned it up a bit and fixed both of these issues:

- Before #40 fix: Extra spaces in front new lines
- After #40 fix: Sometimes spaces would be redacted from the text