Fix: now you can focus on IconButton text by clicking it in the dialog preview
I fudged this up at some point in the past by making an iconButton with no text thinking: “hey now I can’t activate it in the dialog preview”. So I ended up wrapping the whole item in a click event as one does, but this meant you couldn’t click to focus on the [contenteditable] anymore.

In the fix I directed this click event to the image inside the icon button, which would hopefully be better than before.