Fixed AE dockable panel's double title issue. Related to #72
Issue was that AE's dockable SUI takes the titles from the filename and then this window title becomes a sort of secondary title.

If the user would have any reason to use the script as a regular dialog window instead of a dockable panel, make sure the dialog title and the script filename are a match.